+++ src/assets/css/theme.css
@@ -7,8 +7,8 @@
  * Author : LGS1920 Team
  * email: contact@lgs1920.fr
  *
- * Created on: 2025-02-26
- * Last modified: 2025-02-26
+ * Created on: 2025-02-27
+ * Last modified: 2025-02-27
  *
  *
  * Copyright © 2025 LGS1920
@@ -545,6 +545,14 @@ sl-details::part(content) {
     display: flex;
     gap: var(--lgs-gutter-s);
     flex-direction: column;
+}
+
+sl-details svg {
+    margin-right: var(--lgs-gutter-s);
+}
+
+sl-details [slot="prefix"], sl-details [slot="suffix"] {
+    margin-right: 0;
 }
 
 /*******************************************************************************
