+++ src/components/Settings/Panel.jsx
@@ -1,3 +1,19 @@
+/*******************************************************************************
+ *
+ * This file is part of the LGS1920/studio project.
+ *
+ * File: Panel.jsx
+ *
+ * Author : LGS1920 Team
+ * email: contact@lgs1920.fr
+ *
+ * Created on: 2025-02-27
+ * Last modified: 2025-02-27
+ *
+ *
+ * Copyright © 2025 LGS1920
+ ******************************************************************************/
+
 import { SETTINGS_EDITOR_DRAWER }                                     from '@Core/constants'
 import { faPaintbrushPencil, faScrewdriverWrench }                    from '@fortawesome/pro-regular-svg-icons'
 import { faCircleUser }                                               from '@fortawesome/pro-solid-svg-icons'
@@ -7,9 +23,9 @@ import React                                                          from 'reac
 import { useSnapshot }                                                from 'valtio'
 import './style.css'
 import { DrawerFooter }                                               from '../DrawerFooter'
-import { GeneralTools }                                               from './tools/general/GeneralTools'
-import { ProfileTools }                                               from './tools/profile/ProfileTools'
-import { Style }                                                      from './tools/style/Style'
+import { GlobalSettings }                                             from './application/general/GlobalSettings'
+import { ProfileTools }                                               from './application/profile/ProfileTools'
+import { Style }                                                      from './application/style/Style'
 
 export const Panel = () => {
     const drawers = useSnapshot(lgs.mainProxy.drawers)
@@ -32,12 +48,13 @@ export const Panel = () => {
                       contained
                       className={'lgs-theme'}>
                 <SlTabGroup>
+                    <SlTab slot="nav" panel="tab-tools">
+                        <SlIcon library="fa" name={FA2SL.set(faScrewdriverWrench)}/> {'Global Settings'}
+                    </SlTab>
                     <SlTab slot="nav" panel="tab-ui">
                         <SlIcon library="fa" name={FA2SL.set(faPaintbrushPencil)}/> {'User Interface'}
                     </SlTab>
-                    <SlTab slot="nav" panel="tab-tools">
-                        <SlIcon library="fa" name={FA2SL.set(faScrewdriverWrench)}/> {'Tools'}
-                    </SlTab>
+
 
                     <SlTab slot="nav" panel="tab-user" id={'manage-user-profile'}>
                         <SlTooltip placement={'top'} hoist content={'Manage My Profile'}>
@@ -45,7 +62,7 @@ export const Panel = () => {
                         </SlTooltip>
                     </SlTab>
 
-                    <SlTabPanel name="tab-tools"><GeneralTools/></SlTabPanel>
+                    <SlTabPanel name="tab-tools"><GlobalSettings/></SlTabPanel>
                     <SlTabPanel name="tab-ui"><Style/></SlTabPanel>
                     <SlTabPanel name="tab-user" open><ProfileTools/></SlTabPanel>
                 </SlTabGroup>
