+++ src/components/Settings/application/general/GlobalSettings.jsx
@@ -2,29 +2,28 @@
  *
  * This file is part of the LGS1920/studio project.
  *
- * File: Style.jsx
+ * File: GlobalSettings.jsx
  *
  * Author : LGS1920 Team
  * email: contact@lgs1920.fr
  *
- * Created on: 2025-02-26
- * Last modified: 2025-02-26
+ * Created on: 2025-02-27
+ * Last modified: 2025-02-27
  *
  *
  * Copyright © 2025 LGS1920
  ******************************************************************************/
 
-import { MenuSettings }      from '@Components/Settings/tools/style/MenuSettings'
-import { SlDetails }         from '@shoelace-style/shoelace/dist/react'
-import { useEffect, useRef } from 'react'
-import { CameraSettings }    from './CameraSettings'
-import { WelcomeModal }      from './WelcomeModal'
+import { CameraSettings }      from '@Components/Settings/application/general/CameraSettings'
+import { UnitsSystemSettings } from '@Components/Settings/application/general/UnitsSystemSettings'
+import { SlDetails }           from '@shoelace-style/shoelace/dist/react'
+import { useEffect, useRef }   from 'react'
 
-export const Style = () => {
-    const styleSettings = useRef(null)
+export const GlobalSettings = () => {
+    const generalTools = useRef(null)
 
     useEffect(() => {
-        __.ui.ui.initDetailsGroup(styleSettings.current)
+        __.ui.ui.initDetailsGroup(generalTools.current)
     }, [])
 
     const checkClose = (event) => {
@@ -38,14 +37,13 @@ export const Style = () => {
 
     return (
 
-        <div ref={styleSettings} id={'style-settings'}>
-            <SlDetails id={'ui-welcome-modal-settings'}
+        <div ref={generalTools} id={'style-settings'}>
+            <SlDetails id={'tools-unit-system'}
                        small open={false}
                        className={'lgs-theme'}
                        onSlHide={checkClose}
             >
-                <WelcomeModal/>
-
+                <UnitsSystemSettings/>
             </SlDetails>
 
             <SlDetails id={'ui-camera-settings'}
@@ -55,16 +53,6 @@ export const Style = () => {
             >
                 <CameraSettings/>
             </SlDetails>
-
-            <SlDetails id={'ui-menu-settings'}
-                       small open={false}
-                       className={'lgs-theme'}
-                       onSlHide={checkClose}
-            >
-                <MenuSettings/>
-            </SlDetails>
-
-
 
         </div>
