+++ src/components/Settings/application/general/UnitsSystemSettings.jsx
@@ -2,23 +2,25 @@
  *
  * This file is part of the LGS1920/studio project.
  *
- * File: UnitsSystem.jsx
+ * File: UnitsSystemSettings.jsx
  *
  * Author : LGS1920 Team
  * email: contact@lgs1920.fr
  *
- * Created on: 2025-02-26
- * Last modified: 2025-02-26
+ * Created on: 2025-02-27
+ * Last modified: 2025-02-27
  *
  *
  * Copyright © 2025 LGS1920
  ******************************************************************************/
 
+import { faGlobe, faRuler, faRulerCombined }      from '@fortawesome/pro-regular-svg-icons'
+import { FontAwesomeIcon }                        from '@fortawesome/react-fontawesome'
 import { SlDivider, SlRadioButton, SlRadioGroup } from '@shoelace-style/shoelace/dist/react'
 import { DD, DMS, IMPERIAL, INTERNATIONAL }       from '@Utils/UnitUtils'
 import { useSnapshot }                            from 'valtio/index'
 
-export const UnitsSystem = (props) => {
+export const UnitsSystemSettings = (props) => {
 
     const switchValue = (event) => {
         if (window.isOK(event)) {
@@ -38,24 +40,26 @@ export const UnitsSystem = (props) => {
 
     return (
         <>
-            <span slot="summary">{'Units System'}</span>
+            <span slot="summary"><FontAwesomeIcon icon={faRuler}/>{'Units System'}</span>
             <SlDivider/>
             <div className="horizontal-alignment two-columns">
+
                 <SlRadioGroup value={us.current} align-right
                               size={'small'} onSlChange={handleDistanceUnits}
                 >
-                    <label slot="label">{'Distance Units:'}</label>
+                    <label slot="label"><FontAwesomeIcon icon={faRulerCombined}/>{'Distances/Elevations:'}</label>
                     <SlRadioButton value={INTERNATIONAL}>{'Metric'}</SlRadioButton>
                     <SlRadioButton value={IMPERIAL}>{'Impérial'}</SlRadioButton>
                 </SlRadioGroup>
             </div>
+
             <div className="horizontal-alignment two-columns">
                 <SlRadioGroup value={cs.current} align-right
                               size={'small'} onSlChange={handleCoordinateUnits}
                 >
-                    <label slot="label">{'Coordinate Units:'}</label>
-                    <SlRadioButton value={DD}>{'Digital Degres'}</SlRadioButton>
-                    <SlRadioButton value={DMS}>{'Degres Minutes Seconds'}</SlRadioButton>
+                    <label slot="label"><FontAwesomeIcon icon={faGlobe}/>{'Coordinates:'}</label>
+                    <SlRadioButton value={DD}>{'Digital Degrees'}</SlRadioButton>
+                    <SlRadioButton value={DMS}>{'Degrees Minutes Seconds'}</SlRadioButton>
                 </SlRadioGroup>
             </div>
