+++ src/components/Settings/application/profile/ResetProfile.jsx
@@ -1,3 +1,19 @@
+/*******************************************************************************
+ *
+ * This file is part of the LGS1920/studio project.
+ *
+ * File: ResetProfile.jsx
+ *
+ * Author : LGS1920 Team
+ * email: contact@lgs1920.fr
+ *
+ * Created on: 2025-02-27
+ * Last modified: 2025-02-19
+ *
+ *
+ * Copyright © 2025 LGS1920
+ ******************************************************************************/
+
 import { faSquareCheck }  from '@fortawesome/duotone-regular-svg-icons'
 import { faArrowsRotate } from '@fortawesome/pro-regular-svg-icons'
