+++ src/components/Settings/application/style/MenuSettings.jsx
@@ -1,4 +1,20 @@
-import { MenuSample }    from '@Components/Settings/tools/style/MenuSample'
+/*******************************************************************************
+ *
+ * This file is part of the LGS1920/studio project.
+ *
+ * File: MenuSettings.jsx
+ *
+ * Author : LGS1920 Team
+ * email: contact@lgs1920.fr
+ *
+ * Created on: 2025-02-27
+ * Last modified: 2025-02-27
+ *
+ *
+ * Copyright © 2025 LGS1920
+ ******************************************************************************/
+
+import { MenuSample }    from '@Components/Settings/application/style/MenuSample'
 import {
     BOTTOM, DESKTOP_MIN, MENU_BOTTOM_END, MENU_BOTTOM_START, MENU_END_END, MENU_END_START, MENU_START_END,
     MENU_START_START, MOBILE_MAX, START,
